Replace bass slider with numeric control and refine seek slider

+++ BNKaraoke.DJ/ViewModels/DJScreenViewModel.Player.cs
@@ -37,9 +37,9 @@ namespace BNKaraoke.DJ.ViewModels
         private double? _fadeStartTimeSeconds;
         private double? _introMuteSeconds;
 
-        partial void OnBassBoostChanged(double value)
+        partial void OnBassBoostChanged(int value)
         {
-            _videoPlayerWindow?.SetBassGain((float)value);
+            _videoPlayerWindow?.SetBassGain(value);
         }
 
         public void SetWarningMessage(string message)
@@ -475,7 +475,7 @@ namespace BNKaraoke.DJ.ViewModels
                         _videoPlayerWindow = null;
                         return;
                     }
-                    _videoPlayerWindow.SetBassGain((float)BassBoost);
+                    _videoPlayerWindow.SetBassGain(BassBoost);
                     _videoPlayerWindow.SongEnded += VideoPlayerWindow_SongEnded;
                     _videoPlayerWindow.Closed += VideoPlayerWindow_Closed;
                     _videoPlayerWindow.MediaPlayer.PositionChanged += OnVLCPositionChanged;
@@ -666,7 +666,7 @@ namespace BNKaraoke.DJ.ViewModels
                         _videoPlayerWindow = null;
                         return;
                     }
-                    _videoPlayerWindow.SetBassGain((float)BassBoost);
+                    _videoPlayerWindow.SetBassGain(BassBoost);
                     _videoPlayerWindow.SongEnded += VideoPlayerWindow_SongEnded;
                     _videoPlayerWindow.Closed += VideoPlayerWindow_Closed;
                     _videoPlayerWindow.MediaPlayer.PositionChanged += OnVLCPositionChanged;
@@ -1009,7 +1009,7 @@ namespace BNKaraoke.DJ.ViewModels
                         _videoPlayerWindow = null;
                         return;
                     }
-                    _videoPlayerWindow.SetBassGain((float)BassBoost);
+                    _videoPlayerWindow.SetBassGain(BassBoost);
                     _videoPlayerWindow.SongEnded += VideoPlayerWindow_SongEnded;
                     _videoPlayerWindow.Closed += VideoPlayerWindow_Closed;
                     _videoPlayerWindow.MediaPlayer.PositionChanged += OnVLCPositionChanged;
